requirements.txt
mysql-connector-python
saucenao_api
discord.py
discord-components
aiomysql
matplotlib
asyncio
saucenao_api
pygelbooru
